do not use duplicate StringExtensions
