+++ src/OrchardCore/OrchardCore.DisplayManagement/Zones/ZoneShapes.cs
@@ -7,7 +7,7 @@ using OrchardCore.DisplayManagement.Descriptors;
 using OrchardCore.DisplayManagement.Shapes;
 using OrchardCore.DisplayManagement.ViewModels;
 using OrchardCore.Modules;
-using OrchardCore.Mvc.Utilities;
+using OrchardCore.ContentManagement.Utilities;
 
 namespace OrchardCore.DisplayManagement.Zones
 {
